+++ src/social/social.module.ts
@@ -1,7 +1,7 @@
 import { SocialController } from './social.controller';
 import { AppService } from '../app.service';
 import { JwtGoogleStrategy } from '../auth/strategy/jwt-google.strategy';
-import { Module } from '@nestjs/common';
+import { CACHE_MANAGER, Module } from '@nestjs/common';
 import { SocialService } from './social.service';
 import { UserService } from 'src/user/user.service';
 import { AuthService } from '../auth/auth.service';
@@ -11,9 +11,10 @@ import { JwtService } from '@nestjs/jwt';
 import { ProductsEntity } from '../global/entities/products.entity';
 import { ProductImagesEntity } from '../global/entities/productimages.entity';
 import { OrdersEntity } from '../global/entities/orders.entity';
+import { cacheModule } from "../user/user.module";
 
 @Module({
-  imports: [
+  imports: [cacheModule,
     TypeOrmModule.forFeature([
       UserEntity,
       ProductsEntity,
@@ -28,6 +29,7 @@ import { OrdersEntity } from '../global/entities/orders.entity';
     JwtGoogleStrategy,
     UserService,
     AuthService,
+
   ],
 })
 export class SocialModule {}
